Switch to standard server.js
