+++ srv/server.js
@@ -1,22 +1,22 @@
-// load modules
-const express = require('express')
 const cds = require('@sap/cds')
-const helmet = require("helmet")
 const proxy = require('@sap/cds-odata-v2-adapter-proxy')
-// For authentication test
+const helmet = require('helmet')
 const passport = require("passport")
 const xsenv = require("@sap/xsenv")
-xsenv.loadEnv();
-const JWTStrategy = require("@sap/xssec").JWTStrategy
-const services = xsenv.getServices({ xsuaa: { tags: "xsuaa" }})
-const xsuaaCredentials = services.xsuaa
-const jwtStrategy = new JWTStrategy(xsuaaCredentials)
-passport.use(jwtStrategy)
+const swaggerUi = require('swagger-ui-express')
+const swaggerSpec = require('./gen/swagger.json')
 
-// config
-const host = process.env.HOST || '0.0.0.0'
-const port = process.env.PORT || 4004
-const csn = 'gen/csn.json'
+var xsuaaCredentials
+try {
+  xsenv.loadEnv();
+  const JWTStrategy = require("@sap/xssec").JWTStrategy
+  const services = xsenv.getServices({ xsuaa: { tags: "xsuaa" }})
+  xsuaaCredentials = services.xsuaa
+  const jwtStrategy = new JWTStrategy(xsuaaCredentials)
+  passport.use(jwtStrategy)
+} catch (error) {
+  console.warn(error.message)
+}
 
 // Middleware to read JWT sent by client
 function jwtLogger(req, res, next) {
@@ -45,17 +45,18 @@ const readJwt = function(req){
   }
 }
 
-;(async () => {
-  // create new app
-  const app = express()
+cds.on('bootstrap', async (app) => {
   app.use(helmet())
-  app.use(jwtLogger)
-  // Authentication using JWT
-  await app.use(passport.initialize())
-  await app.use(
-    passport.authenticate("JWT", { session: false })
-  )
+  app.use(proxy())
 
+  // Authentication using JWT
+  if(xsuaaCredentials) {
+    app.use(jwtLogger)
+    await app.use(passport.initialize())
+    await app.use(
+      passport.authenticate("JWT", { session: false })
+    )  
+  }
   await app.get("/api/userInfo", function (req, res) {
     if(req.user) {
       res.json(req.user)
@@ -77,40 +78,20 @@ const readJwt = function(req){
       return res.status(403).json({
           error: 'Unauthorized',
           message: `Missing required role: <admin>. Available scopes: ${availableScopes}`
-      });
+      })
     }
-  });
+  })
 
-  //  does ony work with @sap/xssec ^2.2.5
-  await cds.connect('db') // ensure database is connected!
-  // serve odata v4
-  await cds
-    .serve('AdminService')
-    .from(csn)
-    .with('admin-service.js')
-    .in(app)
+  // Swagger / OpenAPI
+  var options = {
+    explorer: true
+  }
+	app.get('/api/api-docs.json', function(req, res) {
+		res.setHeader('Content-Type', 'application/json')
+		res.send(swaggerSpec)
+	})   
+  app.use('/api/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec, options))
 
-  await cds
-    .serve('CatalogService')
-    .from(csn)
-    .with('cat-service.js')
-    .in(app)
+})
 
-  // serve odata v2
-  // process.env.XS_APP_LOG_LEVEL = 'error'; // suppress debug logs
-  app.use(
-    proxy({
-      // app
-      path: 'v2',
-      model: csn,
-      // target
-      port: port
-    })
-  )
-
-  // start server
-  const server = app.listen(port, host, () =>
-    console.info(`app is listing at ${host}:${port}`)
-  )
-  server.on('error', error => console.error(error.stack))
-})()
+module.exports = cds.server // > delegate to default server.js
